Introduce max_count for checking the queue master
If something fails during movement of a queue, the script would hang
in teh same queue awaiting forever.
This patch introduces a max_count option that can be set so if we reach
the number of retries set it skips waiting for the queue to be changed.

This is useful for unnatended launches in clusters where the number of
queues is high and they are heavily unbalanced, as it can take hours to
change masters and sync all of them, the script can be fired and left alone.

alphabetize options