Ignore `sync_queue` errors
If a cluster contains queues that are not mirrored the `sync_queue`
command fails with this error message:

        20210602-00:25:37 [INFO] Synchronising queue 'notifications.sample' in vhost 'openstack'
        Error: not_mirrored

To mitigate such a failure, the script now checks the policy applied
to a particular queue for the `ha-mode` key and skips queues that do
not contain this key.

Note that this error only occurs in the last call to `sync_queue`
after the temporary policy is cleared, exposing the original,
non-mirror policy of the queue. An alternative strategy is to simply
ignore the error.

Signed-off-by: Nicolas Bock <nicolas.bock@canonical.com>